fix(react-native): transform-request-props-example
