+++ packages/native/examples/with-controlled-props/Search.js
@@ -11,7 +11,7 @@ export default function Search({ setResetPagination }) {
       // To fetch suggestions
       searchComponent.triggerDefaultQuery();
     }
-  }, [searchbase, text]);
+  }, [text]);
 
   return (
     <SearchBox
